comments last phase and notifications

+++ CafeteriaApp/CafeteriaApp.Backend/Controllers/Comment.php
@@ -1,4 +1,6 @@
 <?php
+require_once("CafeteriaApp.Backend/Controllers/Dates.php"); 
+
 //result=getCommentsByMenuItemId();
 //if(result)
  // echo json_encode(result);
@@ -15,13 +17,13 @@
   }
   else
   {
-    $sql = "select User.UserName , Comment.Details  from Comment inner join Customer on Comment.CustomerId=Customer.Id inner join User on User.Id=Customer.UserId  where MenuItemId =".$id;
+    $sql = "select User.UserName , Comment.Id ,Comment.Details ,Dates.Date  from Comment inner join User on Comment.UserId=User.Id  inner join Dates on Dates.Id=Comment.DateId  where MenuItemId =".$id;
     $result = $conn->query($sql);
     if ($result)
     {
       $comments = mysqli_fetch_all($result, MYSQLI_ASSOC);
       mysqli_free_result($result);
-        return $comments;   
+        return $comments;
     }
     else
     { //server
@@ -31,7 +33,7 @@
   } 
 }
 
-function getCommentsByCustomerId($conn,$id) // used for editing or deleting comments of a user
+function getCommentsByUserId($conn,$id) // used for editing or deleting comments of a user
 {  
   if (!isset($id)) 
   {
@@ -40,12 +42,12 @@ function getCommentsByCustomerId($conn,$id) // used for editing or deleting comm
   }
   else
   {
-    $sql = "select * from Comment where CustomerId =".$id;
+    $sql = "select * from Comment where UserId =".$id;
     $result = $conn->query($sql);
     if ($result)
     {
-      $comments = mysqli_fetch_assoc($result);
-      mysqli_free_result($result);
+ $comments = mysqli_fetch_all($result, MYSQLI_ASSOC);
+       mysqli_free_result($result);
         return $comments;   
     
     }
@@ -56,10 +58,37 @@ function getCommentsByCustomerId($conn,$id) // used for editing or deleting comm
   } 
 }
 
+function getCommentsIdsByUserIdAndMenuItemId($conn,$cid,$mid) // used for editing or deleting comments of a user
+{  
+  if ( !isset($cid) || !isset($mid)) 
+  {
+    //echo "Error: Id is not set";
+    return 1;
+  }
+  else
+  {
+    $sql = "select Id from Comment where UserId ='".$cid."' and MenuItemId=".$mid ;
+    $result = $conn->query($sql);
+    if ($result)
+    {
+    $comments = mysqli_fetch_all($result, MYSQLI_NUM);
+       mysqli_free_result($result);
+       foreach ($comments as $key => $value) {
+         $comments[$key]= ($value[0]) ;
+       }
+        return $comments;   
+    
+    }
+    else
+    {
+      echo "Error retrieving Comments: " . $conn->error;
+    }
+  } 
+}
 
 function addComment($conn,$details ,$Cid,$Mid)
 {
-  if (!isset($details)) 
+  if (!isset($details))
   {
     //echo "Error: Comment Details is not set";
     return;
@@ -76,15 +105,19 @@ function addComment($conn,$details ,$Cid,$Mid)
   }
   else
   {
-    $sql = "insert into Comment (Details , CustomerId ,MenuItemId ) values (?,?,?)";
+
+    $DateId =getDateIdByDate($conn,date("Y-m-d"));
+
+    $sql = "insert into Comment (Details , UserId , MenuItemId , DateId ) values (?,?,?,?)";
     $stmt = $conn->prepare($sql);
-    $stmt->bind_param("sii",$Details, $CustomerId ,$MenuItemId);
+    $stmt->bind_param("siii",$Details, $UserId ,$MenuItemId,$DateId);
     $Details = $details;
-    $CustomerId = $Cid;
+    $UserId = $Cid;
     $MenuItemId=$Mid;
     if ($stmt->execute()===TRUE)
     {
-      return "Comment Added successfully";
+       return  mysqli_insert_id($conn);
+      //return "Comment Added successfully";
     }
     else
     {
@@ -93,7 +126,7 @@ function addComment($conn,$details ,$Cid,$Mid)
   }
 }
 
-function editComment($conn,$details,$id)
+function editComment($conn,$details,$id)// check the customer if he owns the comment
 {
   if (!isset($details)) 
   {
@@ -107,9 +140,10 @@ function editComment($conn,$details,$id)
   }
   else
   {
-    $sql = "update Comment set Details = (?) where Id = (?)";
+    $DateId =getDateIdByDate($conn,date("Y-m-d"));
+    $sql = "update Comment set Details = (?) , DateId= (?) where Id = (?)";
     $stmt = $conn->prepare($sql);
-    $stmt->bind_param("si",$Details,$Id);
+    $stmt->bind_param("sii",$Details,$DateId,$Id);
     $Details = $details;
     $Id = $id;
     if ($stmt->execute()===TRUE)
@@ -123,8 +157,7 @@ function editComment($conn,$details,$id)
   } 
 }
 
-
-function deleteComment($conn,$id)
+function checkOwnershipOfComment($conn,$id,$cid)//check if its for the customer before deleting
 {
   if (!isset($id))
   {
@@ -133,7 +166,33 @@ function deleteComment($conn,$id)
   }
   else
   {
-    $conn->query("set foreign_key_checks=0");
+    $sql = "select count(*) from Comment where Id = ".$id. " and UserId=".$cid;
+    $result = $conn->query($sql);
+    if ($result) {
+      $count= mysqli_fetch_array($result , MYSQLI_NUM);
+      $count=(int)$count[0];
+      if ($count===1)
+    {
+      return true;
+    }
+    else
+    {
+      return false ;
+    }
+    }
+   
+  } 
+}
+
+function deleteComment($conn,$id)//check if its for the customer before deleting
+{
+  if (!isset($id))
+  {
+    //echo "Error: Id is not set";
+    return;
+  }
+  else
+  {
     $sql = "delete from Comment where Id = ".$id. " LIMIT 1";
     if ($conn->query($sql)===TRUE)
     {
